Move failures search view
