+++ frontend/src/views/failures/search/index.scss
@@ -1,0 +1,24 @@
+.failures-search {
+  max-width: 1020px;
+  min-height: 100%;
+
+  background-color: #e4e4e4;
+  margin: 0 auto;
+  padding: 20px 0;
+
+  .page-loader {
+    text-align: center;
+    padding: 50px 0;
+  }
+
+  .empty-event {
+    margin: 50px 15px 0;
+    font-size: 30px;
+    font-weight: lighter;
+    font-family: Roboto;
+  }
+
+  & > .event {
+    margin: 10px 10px;
+  }
+}
